add an in-memory cache
Processing large data sets with frequent duplication performs better
when we use an in-memory cache to supplement the on-disk cache. The
lru_cache decorator in Python requires all arguments to the decorated
function to be hashable, so this patch converts the existing factory
functions to classes so the fetch() methods can be decorated and still
have access to the underlying cache object.

Change-Id: I1144874e76b5db8fd125679da3a4b5b9ec4540cd
Signed-off-by: Doug Hellmann <doug@doughellmann.com>