add --highlight-sponsors option to contributor report commands
When the option is specified contributions from organizations that are
not sponsors are combined into a single "*other" result to preserve
the overall relative contribution levels of the remaining companies
but to limit the number of names returned in the output.

The default is to show all sponsors, but a --sponsor-level option can
change that to specify only platinum or gold members.

Change-Id: I3062dc607ee2f35a73cfc2237a8f9c58df78af8e
Signed-off-by: Doug Hellmann <doug@doughellmann.com>