be more forgiving of bad affiliation date ranges
Scan all affiliations and take the latest one that has a start date
that makes it valid for the date in question. This works around the
fact that there is a bug in the web UI for the foundation site making
it difficult for users to modify their existing affiliations.

Change-Id: I369d777e0ef2064fc728f9042d9b0e3eca84d350
Signed-off-by: Doug Hellmann <doug@doughellmann.com>