improve error handling
Use error code 2 to indicate that we are exiting a program without
doing any work. This leaves error code 1 to mean that the program
actually had an error, and we can check for those errors and report
them.

Add --debug and -v flag where missing in calls to python3-first app.

Change-Id: Ia6f39dc8eee3505ac573670a9e44ae225ed48334
Signed-off-by: Doug Hellmann <doug@doughellmann.com>